+++ src/main/java/com/mateusz/springgpt/repository/HeatmapRepository.java
@@ -1,9 +1,8 @@
 package com.mateusz.springgpt.repository;
 
-import com.mateusz.springgpt.entity.Heatmap;
+import com.mateusz.springgpt.entity.HeatmapEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ScreenshotRepository extends JpaRepository<Heatmap, Long> {
-}
+public interface HeatmapRepository extends JpaRepository<HeatmapEntity, Long> {}