+++ src/main/java/com/mateusz/springgpt/service/HeatmapAnalysisService.java
@@ -1,0 +1,37 @@
+package com.mateusz.springgpt.service;
+
+import com.mateusz.springgpt.entity.HeatmapEntity;
+import com.mateusz.springgpt.repository.HeatmapRepository;
+import com.mateusz.springgpt.service.tools.ImageAnalyzer;
+import lombok.extern.slf4j.Slf4j;
+import org.bytedeco.opencv.opencv_core.Mat;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+
+@Service
+@Slf4j
+public class HeatmapAnalysisService {
+
+    private final HeatmapRepository heatmapRepository;
+
+    @Autowired
+    public HeatmapAnalysisService(HeatmapRepository heatmapRepository) {
+        this.heatmapRepository = heatmapRepository;
+    }
+
+    public Double analyzeHeatmap(Long id) {
+        Optional<HeatmapEntity> heatmap = heatmapRepository.findById(id);
+
+        if (heatmap.isEmpty()) {
+            return null;
+        }
+
+        String base64Image = heatmap.get().getBase64();
+        log.info("Analyzing pixels of heatmap number: {}", id);
+        Mat image = ImageAnalyzer.base64ToMat(base64Image);
+
+        return ImageAnalyzer.greenRedRatio(image);
+    }
+}